js: update paths after rename
